Clean-up this mofo!
