--- app/src/layout.jsx
@@ -1,7 +1,0 @@
-// @flow
-
-import React from 'react'
-import { NavLink } from '$shared/components/MobileNav'
-import Link from './marketplace/components/Link'
-
-NavLink.Link = <Link />
