+++ app/src/marketplace/components/Layout/layout.pcss
@@ -23,10 +23,6 @@
   padding-right: 1.5rem;
 }
 
-.desktopNav {
-  display: none;
-}
-
 @media (--sm-down) {
   .desktopNav {
     display: none;
@@ -35,6 +31,6 @@
 
 @media (--md-up) {
   .mobileNav {
-    /* display: none; */
+    display: none;
   }
 }
