--- app/src/marketplace/components/Nav/nav.pcss
@@ -1,38 +1,0 @@
-.accountCircle {
-  float: right;
-  margin: -4px 15px -4px 0;
-}
-
-/* Using tag name (a) for higher specificity. Not that !important. */
-a.sansLink {
-  cursor: pointer;
-  font-family: var(--sans);
-  font-size: calc(0.875 * var(--um));
-  font-weight: var(--regular);
-  letter-spacing: 0;
-  line-height: 2em;
-  padding: 0 var(--um);
-  text-transform: none;
-}
-
-.userInfo {
-  color: #323232;
-  font-size: calc(0.875 * var(--um));
-  font-family: var(--sans);
-  margin: calc(0.25 * var(--um)) 0;
-  padding: 0 var(--um) calc(0.375 * var(--um));
-  text-transform: none;
-}
-
-.userUsername {
-  color: #525252;
-  font-size: calc(0.75 * var(--um));
-  font-weight: var(--regular);
-}
-
-@media (--md-up) {
-  .accountCircle {
-    float: none;
-    margin: 0;
-  }
-}
