--- app/src/marketplace/containers/Nav/index.jsx
@@ -1,16 +1,0 @@
-// @flow
-
-import { connect } from 'react-redux'
-import Nav from '../../components/Nav'
-import { selectUserData } from '$shared/modules/user/selectors'
-import type { User } from '$shared/flowtype/user-types'
-
-type StateProps = {
-    currentUser: ?User,
-}
-
-const mapStateToProps = (state): StateProps => ({
-    currentUser: selectUserData(state),
-})
-
-export default connect(mapStateToProps)(Nav)
