+++ app/src/shared/components/MobileNav/LogoItem/logoItem2.pcss
