--- app/src/shared/components/MobileNav/NavDivider/index.jsx
@@ -1,9 +1,0 @@
-// @flow
-
-import React from 'react'
-import classNames from 'classnames'
-import styles from './navDivider.pcss'
-
-const NavDivider = () => <li className={classNames(styles.divider, 'd-md-none')} />
-
-export default NavDivider
