--- app/src/shared/components/MobileNav/NavDivider/navDivider.pcss
@@ -1,17 +1,0 @@
-.divider {
-  clear: both;
-  padding: 1em 15px;
-
-  &::after {
-    background: #EAEAEA;
-    content: '';
-    display: block;
-    height: 1px;
-  }
-}
-
-@media (--md-up) {
-  .divider + li {
-    margin-left: 1.5em;
-  }
-}
