--- app/src/shared/components/MobileNav/NavDropdown/index.jsx
@@ -1,52 +1,0 @@
-// @flow
-
-import * as React from 'react'
-import classNames from 'classnames'
-
-import NavLink from '../NavLink'
-import navLinkStyles from '../NavLink/navLink.pcss'
-import styles from './navDropdown.pcss'
-
-type Props = {
-    label?: React.Node,
-    toggle?: React.Node,
-    children: React.Node,
-    align?: string,
-    noPointer?: boolean,
-}
-
-const NavDropdown = ({
-    label,
-    toggle,
-    children,
-    align,
-    noPointer,
-    ...props
-}: Props) => (
-    <div className={classNames(styles.dropdown, 'd-none', 'd-md-inline-block', navLinkStyles.navLinkParent)}>
-        {toggle || (
-            <NavLink {...props}>
-                {label || null}
-            </NavLink>
-        )}
-        <div
-            className={classNames(styles.dropdownMenuWrapper, {
-                [styles.centered]: !align || align === 'center',
-                [styles.pullLeft]: align === 'left',
-                [styles.pullRight]: align === 'right',
-            })}
-        >
-            <ul
-                className={styles.dropdownMenu}
-            >
-                {React.Children.map(children, (child) => (child ? (
-                    <li>{child}</li>
-                ) : (
-                    <li className={styles.separator} />
-                )))}
-            </ul>
-        </div>
-    </div>
-)
-
-export default NavDropdown
