--- app/src/shared/components/MobileNav/NavDropdown/navDropdown.pcss
@@ -1,87 +1,0 @@
-.dropdownMenuWrapper {
-  left: 50%;
-  min-width: 102%;
-  opacity: 0;
-  padding-top: 0;
-  position: absolute;
-  transition: 200ms ease-out;
-  transition-delay: 200ms, 0s, 0s;
-  transition-property: visibility, opacity, padding-top;
-  visibility: hidden;
-  z-index: 3;
-}
-
-.pullLeft {
-  transform: translateX(-80%);
-}
-
-.pullRight {
-  left: 0;
-  transform: translateX(calc(-1 * var(--um)));
-}
-
-.centered {
-  transform: translateX(-50%);
-}
-
-.dropdownMenu {
-  background: #FFFFFF;
-  border-radius: calc(0.25 * var(--um));
-  box-shadow: 0 0 calc(0.375 * var(--um)) rgba(0, 0, 0, 0.1);
-  font-size: calc(0.75 * var(--um));
-  line-height: 1em;
-  list-style: none;
-  margin: 0;
-  min-width: 9em;
-  padding: calc(0.25 * var(--um)) 0;
-  position: relative;
-
-  li {
-    display: block;
-    text-transform: uppercase;
-  }
-
-  a {
-    background-color: #FFFFFF;
-    display: block;
-    letter-spacing: calc(0.125 * var(--um));
-    line-height: calc(1.75 * var(--um));
-    padding: 0 var(--um);
-    transition: 200ms ease-in-out background-color;
-    white-space: nowrap;
-
-    &,
-    &:--idle,
-    &:--enter {
-      color: #323232;
-      text-decoration: none;
-    }
-
-    &:--enter {
-      background-color: #F8F8F8;
-      transition-duration: 50ms;
-    }
-
-    &.active {
-      cursor: default;
-    }
-  }
-}
-
-.dropdown {
-  outline: 0;
-  position: relative;
-
-  &:--enter .dropdownMenuWrapper {
-    opacity: 1;
-    padding-top: 5px;
-    transition-delay: 0s, 0s, 0s;
-    visibility: visible;
-  }
-}
-
-.separator {
-  background-color: #EFEFEF;
-  height: 1px;
-  margin: calc(0.25 * var(--um)) 0;
-}
