--- app/src/shared/components/MobileNav/NavHamburger/navHamburger.pcss
@@ -1,40 +1,0 @@
-.link {
-  display: block;
-  float: right;
-  height: 24px;
-  position: relative;
-  width: 24px;
-  z-index: 2;
-
-  svg {
-    display: block;
-    left: 2px;
-    position: absolute;
-    top: 5px;
-  }
-}
-
-.x,
-.hamburger {
-  transition: visibility 0s, opacity 300ms;
-}
-
-.open .x,
-.closed .hamburger {
-  opacity: 1;
-  visibility: visible;
-}
-
-.closed .x,
-.open .hamburger {
-  opacity: 0;
-  transition-delay: 300ms, 0s;
-  visibility: hidden;
-}
-
-@media (--md-up) {
-  .wrapper,
-  .link {
-    display: none;
-  }
-}
