--- app/src/shared/components/MobileNav/NavItem/navItem.pcss
@@ -1,20 +1,0 @@
-@media (--sm-down) {
-  .item {
-    float: left;
-    width: 50%;
-  }
-}
-
-@media (--md-up) {
-  .item {
-    padding: 0;
-  }
-
-  .item + .item {
-    margin-left: 1.5em;
-  }
-
-  .outline {
-    padding: 0.25em 0;
-  }
-}
