--- app/src/shared/components/MobileNav/NavLabel/navLabel.pcss
@@ -1,14 +1,0 @@
-.label {
-  clear: both;
-  color: #848484;
-  font-size: 0.75em;
-  font-weight: var(--medium);
-  padding: 1em 15px;
-  text-transform: uppercase;
-}
-
-@media (--md-up) {
-  .divider + li {
-    margin-left: 1.5em;
-  }
-}
