--- app/src/shared/components/MobileNav/NavLink/navLink.pcss
@@ -1,97 +1,0 @@
-.navLink {
-  border-radius: 2px;
-  display: block;
-  font-size: calc(0.75 * var(--um));
-  letter-spacing: 2px;
-  line-height: 2.5em;
-  overflow: hidden;
-  padding: 0 15px;
-  text-overflow: ellipsis;
-  text-transform: uppercase;
-  transition: 200ms ease-in-out color;
-  vertical-align: top;
-  white-space: nowrap;
-
-  &,
-  &:--enter,
-  &:--idle {
-    color: #323232;
-    outline: none;
-    text-decoration: none;
-  }
-
-  &:--enter {
-    color: #0324FF;
-    transition-duration: 50ms;
-  }
-}
-
-.outline {
-  display: inline-block;
-  line-height: 2.375em;
-  margin-left: 15px;
-  max-width: 75%;
-}
-
-@media (--sm-down) {
-  .navLink.outline {
-    &,
-    &:--enter,
-    &:--idle {
-      background: transparent;
-      border-color: #323232;
-      color: #323232;
-    }
-  }
-}
-
-@media (--md-up) {
-  .inner {
-    font-size: 0.75em;
-    position: relative;
-  }
-
-  .navLinkParent:--enter {
-    background: transparent;
-  }
-
-  .navLink {
-    display: block;
-    max-width: none;
-    padding: 0;
-
-    &,
-    &:--enter,
-    &:--idle {
-      color: white;
-    }
-
-    &:--enter {
-      background: transparent;
-    }
-  }
-
-  .outline {
-    line-height: 1.875em;
-    margin-left: 0;
-    padding: 0 1em;
-  }
-
-  .opaqueNav {
-    &,
-    &:--enter,
-    &:--idle {
-      color: #323232;
-    }
-
-    &.outline {
-      &,
-      &:--enter,
-      &:--idle {
-        background: transparent;
-        border-color: #323232;
-        color: #323232;
-      }
-    }
-  }
-}
