--- app/src/shared/components/MobileNav/NavLogo/navLogo.pcss
@@ -1,20 +1,0 @@
-.logoLink {
-  display: block;
-  float: left;
-  height: 1.5em;
-  outline: 0;
-  width: 1.5em;
-}
-
-.logo {
-  display: block;
-  height: 100%;
-  width: 100%;
-}
-
-@media (--md-up) {
-  .logoLink {
-    height: 2.5em;
-    width: 2.5em;
-  }
-}
