--- app/src/shared/components/MobileNav/nav.pcss
@@ -1,137 +1,0 @@
-.nav {
-  font-family: 'IBM Plex Mono', 'Hiragino Sans GB', '冬青黑体', 'SimHei', '黑体', sans-serif;
-  font-size: 16px;
-  font-weight: var(--semiBold);
-  line-height: 1em;
-}
-
-.overlay {
-  left: 0;
-  position: absolute;
-  top: 0;
-  width: 100%;
-  z-index: 10;
-}
-
-.nav,
-.toggleWrapper {
-  padding-bottom: 1.25em;
-  padding-top: 1.25em;
-}
-
-.noNav .nav {
-  display: none;
-}
-
-.inner {
-  position: relative;
-}
-
-.inner,
-.toggleWrapper,
-.navItems {
-  &::after {
-    clear: both;
-    content: '';
-    display: block;
-  }
-}
-
-.opaque {
-  background: white;
-}
-
-.navItemsWrapper {
-  background: white;
-  height: 100%;
-  left: 0;
-  opacity: 0;
-  padding: 0 0 2.5em;
-  position: fixed;
-  top: 0;
-  transition: visibility 0s 300ms, opacity 300ms ease-in-out;
-  visibility: hidden;
-  width: 100%;
-  z-index: 10;
-}
-
-.navItems {
-  list-style: none;
-  margin: 0;
-  padding: 0.75em 15px 0;
-  transition: padding-top 300ms ease-in-out;
-}
-
-.open {
-  .navItems {
-    padding-top: 0;
-  }
-
-  .navItemsWrapper {
-    opacity: 1;
-    overflow: auto;
-    transition-delay: 0s, 0s;
-    visibility: visible;
-  }
-}
-
-.fullWidth :global(.container) {
-  width: auto;
-  max-width: 100%;
-}
-
-.marketplaceLabel {
-  color: var(--greyDark);
-  float: left;
-  font-size: 0.75em;
-  font-weight: var(--medium);
-  line-height: 2em;
-  margin-left: 1.25em;
-  text-transform: uppercase;
-}
-
-@media (--sm-down) {
-  .nav :global(.container) {
-    padding-left: 30px;
-    padding-right: 30px;
-  }
-}
-
-@media (--md-up) {
-  .nav {
-    font-weight: var(--medium);
-  }
-
-  .fullWidth :global(.container) {
-    padding-left: 30px;
-    padding-right: 30px;
-  }
-
-  .navItems {
-    align-items: center;
-    display: flex;
-    float: right;
-    margin: 0;
-    padding: 0;
-    transition: none;
-  }
-
-  .toggleWrapper {
-    display: none;
-  }
-
-  .navItemsWrapper {
-    background: transparent;
-    float: right;
-    height: auto;
-    opacity: 1;
-    padding: 0;
-    position: initial;
-    visibility: visible;
-    width: auto;
-  }
-
-  .open .navItemsWrapper {
-    overflow: visible;
-  }
-}
